requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
textblob>=0.17.0
wordcloud>=1.9.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-magic>=0.4.27